% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_zoi_local.R
\name{calc_zoi_local}
\alias{calc_zoi_local}
\title{Correct the ZOI after local changes/perturbations}
\usage{
calc_zoi_local(
  vects,
  ref_raster,
  mod,
  scenario_table,
  s_gid = 1,
  model = c("suit", "perm")[1],
  baseline = FALSE,
  where = c("R", "GRASS")[1],
  cross_chr = "cross",
  separator = "_",
  verbose = FALSE,
  g_module = c("r.mfilter", "r.resamp.filter", "r.neighbors")[2],
  output_type = c("cumulative_zoi", "density")[1]
)
}
\description{
Describe function here.
}
