% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_along.R
\name{extract_along}
\alias{extract_along}
\title{Extract values of raster using in parallel}
\usage{
extract_along(
  r,
  sl,
  ws,
  col_step_length = "dist",
  step_id = "use_ava_data_animals_id",
  prefix = "along_"
)
}
\arguments{
\item{r}{\verb{[SpatRast]} \cr Set of rasters, in the \code{SpatRaster} format from \link{terra}.}

\item{sl}{Vector of step lines, with a column representing the step id
(stratum or just step number) and a LINESTRING geometry for each step.
Can be a \code{sf} or a \code{SpatVector} object.}

\item{ws}{List of which summary statistics are to be computed. Currently, only "mean", "max",
"min", and "sum" are implemented.}
}
\description{
Extract values of raster using in parallel
}
