% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_zoi.R
\name{model_get_zoi}
\alias{model_get_zoi}
\title{Get ZOI radius from covariates in a model}
\usage{
model_get_zoi(mod, variable, season, model, separator = "_")
}
\arguments{
\item{mod}{Model.}

\item{variable}{Vector of variable names or parts of variable names, to be used
to select them from the model formula.}
}
\description{
Get ZOI radius from covariates in a model
}
