% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_extract.R
\name{parallel_extract}
\alias{parallel_extract}
\title{Extract values of raster using in parallel}
\usage{
parallel_extract(
  r,
  sl,
  ws,
  step_id = "use_ava_data_animals_id",
  prefix = "along_"
)
}
\arguments{
\item{r}{Stack of \code{SpatRaster} objects}

\item{sl}{Vector of step lines, with a column representing the step id
(stratum) and a LINESTRING geometry for each step. Can be an \code{sf} or a
\code{SpatVector} object.}

\item{ws}{List of}
}
\description{
Extract values of raster using in parallel
}
