% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_plot.R
\name{scenario_plot_vect}
\alias{scenario_plot_vect}
\title{Plot vectors to used in a scenario}
\usage{
scenario_plot_vect(
  vects,
  add = TRUE,
  col_pts = "red",
  col_lin = "black",
  col_pol = "grey",
  pal_pts = "cat",
  pal_lin = "div",
  pal_pol = "div",
  ...
)
}
\arguments{
\item{vects}{List of vectors with points, lines, and polygons to be changed in
one or more scenarios.}

\item{add}{Add all plots in a single one? Only that is implemented. Alternatively,
it could be facets.}

\item{col_pts}{point colors

Ideas: Add subset to conditions (only change == "remove")
Add all features of a scenario set, and highlight only the ones in a scenario}
}
\description{
Plot vectors to used in a scenario
}
