% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_info.R
\name{scenario_table}
\alias{scenario_table}
\alias{scenario_table_create}
\alias{scenario_table_update}
\alias{scenario_table_add}
\alias{get_max_s_gid}
\title{Create a new scenario table}
\usage{
scenario_table_create(
  s_gid = 1,
  scenset_id = NA,
  scenset_name = NA,
  scen_id = NA,
  scen_name = NA,
  scen_nickname = NA,
  project_name = NA,
  study_area = NA,
  sub_area = NA,
  species = NA,
  season = NA,
  intra_inter_season = c("intra", "inter", "all")[1],
  baseline = NA,
  combination = NA,
  description = NA,
  region_grass_vector = NA,
  region_grass_raster = NA,
  region_file_vector = NA,
  region_file_raster = NA,
  region_postgis_layer = NA,
  region_postgis_condition = NA,
  grass_mapset = NA,
  model_suit_object_path = NA,
  model_suit_object = NA,
  model_suit_formula = NA,
  model_suit_covariates = NA,
  model_perm_object_path = NA,
  model_perm_object = NA,
  model_perm_formula = NA,
  model_perm_covariates = NA,
  data_suit_object_path = NA,
  data_suit_object = NA,
  data_perm_object_path = NA,
  data_perm_object = NA,
  data_env_object_path = NA,
  data_env_object = NA,
  modified_vector_pts = NA,
  modified_vector_lin = NA,
  modified_vector_pol = NA,
  modified_raster = NA,
  modified_env_layers = NA,
  modified_var_names_suit = NA,
  modified_var_names_perm = NA,
  zoi_radii_suit = NA,
  zoi_radii_perm = NA,
  baseline_max_suit = NA,
  baseline_max_perm = NA,
  scenario_folder_path = NA,
  scenario_map_prefix = NA
)

scenario_table_update(
  scenario_table,
  s_gid = NA,
  ref_column = c("s_gid", "scen_id")[1],
  scenset_id = NA,
  scenset_name = NA,
  scen_id = NA,
  scen_name = NA,
  scen_nickname = NA,
  project_name = NA,
  study_area = NA,
  sub_area = NA,
  species = NA,
  season = NA,
  intra_inter_season = NA,
  baseline = NA,
  combination = NA,
  description = NA,
  region_grass_vector = NA,
  region_grass_raster = NA,
  region_file_vector = NA,
  region_file_raster = NA,
  region_postgis_layer = NA,
  region_postgis_condition = NA,
  grass_mapset = NA,
  model_suit_object_path = NA,
  model_suit_object = NA,
  model_suit_formula = NA,
  model_suit_covariates = NA,
  model_perm_object_path = NA,
  model_perm_object = NA,
  model_perm_formula = NA,
  model_perm_covariates = NA,
  data_suit_object_path = NA,
  data_suit_object = NA,
  data_perm_object_path = NA,
  data_perm_object = NA,
  data_env_object_path = NA,
  data_env_object = NA,
  modified_vector_pts = NA,
  modified_vector_lin = NA,
  modified_vector_pol = NA,
  modified_raster = NA,
  modified_env_layers = NA,
  modified_var_names_suit = NA,
  modified_var_names_perm = NA,
  zoi_radii_suit = NA,
  zoi_radii_perm = NA,
  baseline_max_suit = NA,
  baseline_max_perm = NA,
  scenario_folder_path = NA,
  scenario_map_prefix = NA,
  overwrite = FALSE
)

scenario_table_add(
  scenario_table,
  s_gid = get_max_s_gid(scenario_table) + 1,
  scenset_id = NA,
  scenset_name = NA,
  scen_id = NA,
  scen_name = NA,
  scen_nickname = NA,
  project_name = NA,
  study_area = NA,
  sub_area = NA,
  species = NA,
  season = NA,
  intra_inter_season = c("intra", "inter", "all")[1],
  baseline = NA,
  combination = NA,
  description = NA,
  region_grass_vector = NA,
  region_grass_raster = NA,
  region_file_vector = NA,
  region_file_raster = NA,
  region_postgis_layer = NA,
  region_postgis_condition = NA,
  grass_mapset = NA,
  model_suit_object_path = NA,
  model_suit_object = NA,
  model_suit_formula = NA,
  model_suit_covariates = NA,
  model_perm_object_path = NA,
  model_perm_object = NA,
  model_perm_formula = NA,
  model_perm_covariates = NA,
  data_suit_object_path = NA,
  data_suit_object = NA,
  data_perm_object_path = NA,
  data_perm_object = NA,
  data_env_object_path = NA,
  data_env_object = NA,
  modified_vector_pts = NA,
  modified_vector_lin = NA,
  modified_vector_pol = NA,
  modified_raster = NA,
  modified_env_layers = NA,
  modified_var_names_suit = NA,
  modified_var_names_perm = NA,
  zoi_radii_suit = NA,
  zoi_radii_perm = NA,
  baseline_max_suit = NA,
  baseline_max_perm = NA,
  scenario_folder_path = NA,
  scenario_map_prefix = NA
)

get_max_s_gid(scenario_table)
}
\arguments{
\item{s_gid}{\verb{[numeric(1)]} \cr Scenario general ID - a numeric ID for the
scenario. It should be unique between all scenarios and scenario sets.}

\item{scenset_id}{\verb{[numeric(1)=NA]} \cr Scenario set ID - a numeric ID for the
scenario set.}

\item{scenset_name}{\verb{[character(1)=NA]} \cr Scenario set name. A character name
for the scenario set. It should be as informative as possible (e.g. including
the project name, study area, sub area, species, what is done in the scenario
set).}

\item{scen_id}{\verb{[numeric(1)=NA]} \cr Scenario ID. It should be unique within a
scenario set, but in contrast to s_gid, can be repeated in different scenario
sets.}

\item{scen_name}{\verb{[character(1)=NA]} \cr Scenario name. A character name for the
scenario. It should be as informative as possible (e.g. including the project
name, study area, sub area, species, what is done in the scenario).}

\item{scen_nickname}{\verb{[character(1)=NA]} \cr Scenario "nickname". A character
name for the scenario which is shorter than the actual name and easier to
refer to.}

\item{project_name}{\verb{[character(1)=NA]} \cr Name of the project.}

\item{study_area}{\verb{[character(1)=NA]} \cr Name of the study area.}

\item{sub_area}{\verb{[character(1)=NA]} \cr Name of the subarea, i.e. a spatial
reference within the study area where the proposed local changes are made.}

\item{species}{\verb{[character(1)=NA]} \cr Species name.}

\item{season}{\verb{[character(1)=NA]} \cr Season(s). It can be either one season
if \code{intra_inter_season = "intra"} or two seasons if \code{intra_inter_season = "inter"}
or \code{"all"}.}

\item{intra_inter_season}{\verb{[character(1)="inter"]}{"intra", "inter", "all"} \cr
Whether the scenario is looking at connectivity within season (\code{"intra"}),
between seasons (\code{"inter"}), or both.}

\item{description}{\verb{[character(1)=NA]} \cr Textual description of the model.}

\item{region_grass_vector}{\verb{[character(1)=NA]} \cr Polygon in GRASS used to define
the region and mask. Should be in the format \code{"vector_name@mapset_name"}.}

\item{region_grass_raster}{\verb{[character(1)=NA]} \cr Raster in GRASS used to define
the resolution and alignments of the region, and possible the extent as well.
Should be in the format \code{"raster_name@mapset_name"}.}

\item{region_file_vector}{\verb{[character(1)=NA]} \cr Full path to polygon used to
define the region and mask. Should be in the format \code{"folder_name/file_name.extension"}.}

\item{region_file_raster}{\verb{[character(1)=NA]} \cr Raster used to define
the resolution and alignments of the region, and possible the extent as well.
Should be in the format \code{"folder_name/file_name.extension"}.}

\item{region_postgis_condition}{= NA}

\item{model_suit_object_path}{\verb{[character(1)=NA]} \cr Path to the habitat
suitability model (RSF/ENM) object with coefficient estimates.}

\item{model_suit_object}{\verb{[character(1)=NA]} \cr Habitat suitability model
object (RSF/ENM) with coefficient estimates.}

\item{model_suit_formula}{\verb{[character(1)=NA]} \cr Formula for the habitat
suitability model (RSF/ENM).}

\item{model_suit_covariates}{\verb{[character(1)=NA]} \cr Covariates included in the
habitat suitability model (RSF/ENM).}

\item{model_perm_object_path}{\verb{[character(1)=NA]} \cr Path to the habitat
permeability model (SSF) object with coefficient estimates.}

\item{model_perm_object}{\verb{[character(1)=NA]} \cr Habitat permeability model
object (SSF) with coefficient estimates.}

\item{model_perm_formula}{\verb{[character(1)=NA]} \cr Formula for the habitat
permeability model (SSF).}

\item{model_perm_covariates}{\verb{[character(1)=NA]} \cr Covariates included in the
habitat permeability model (SSF).}

\item{data_suit_object_path}{\verb{[character(1)=NA]} \cr}

\item{data_suit_object}{\verb{[character(1)=NA]} \cr}

\item{data_perm_object_path}{\verb{[character(1)=NA]} \cr}

\item{data_perm_object}{\verb{[character(1)=NA]} \cr}

\item{data_env_object_path}{\verb{[character(1)=NA]} \cr}

\item{data_env_object}{\verb{[character(1)=NA]} \cr}

\item{modified_vector_pts}{\verb{[character(1)=NA]} \cr}

\item{modified_vector_lin}{\verb{[character(1)=NA]} \cr}

\item{modified_vector_pol}{\verb{[character(1)=NA]} \cr}

\item{modified_raster}{\verb{[character(1)=NA]} \cr}

\item{modified_env_layers}{\verb{[character(1)=NA]} \cr}

\item{scenario_table}{\verb{[data.frame]} \cr Table of scenarios (as created
through the \code{scenario_table_create()} function).}

\item{ref_column}{\verb{[character(1)="s_gid"]\{"s_gid", "scen_id"\}} \cr
Column of the scenario table \code{scenario_table} to be used as a reference to select
the rows to be updated. Default is column \code{"s_gid"}.}

\item{overwrite}{\verb{[logical(1)=FALSE]} \cr Should a column value be replaced?
Default is \code{FALSE}.}

\item{region_postgis_table}{\verb{[character(1)=NA]} \cr Polygon in PostGIS used to define
the region and mas. Should be in the format \code{"schema.table"}.}

\item{region_file_path}{Raster or vector to define the region}

\item{region_file_mask}{Raster or vector to use as mask}

\item{zoi_variables}{\verb{[character(1)=NA]} \cr}

\item{zoi_radii}{\verb{[character(1)=NA]} \cr}
}
\description{
Functions to create and manage tables to organize and describe scenarios and
sets of scenarios in R.

\code{\link[=scenario_table_create]{scenario_table_create()}} creates a table to organize and
describe scenarios and sets of scenarios.
If no parameters are given, the table is initialized with NA values, which
can be updated later using \code{\link[=scenario_table_update]{scenario_table_update()}}.

\code{\link[=scenario_table_update]{scenario_table_update()}} updates a table of scenarios, filling
or replacing values of existing rows/scenarios in the table. The reference to
the scenario(s) to be updated is given by the scenario ID(s)
(\code{s_gid} or \code{scen_id} columns) and is defined by the \code{ref_column} parameter.
If the scenario already has a non-null value for one of the columns to be
updated, the \code{overwrite} option might be used to replace the existing values.

\code{\link[=scenario_table_add]{scenario_table_add()}} adds new lines/scenarios to an existing
table of scenarios.

\code{\link[=get_max_s_gid]{get_max_s_gid()}} returns the maximum scenario gid (s_gid) of
an exiting table of scenarios.
}
\details{
The parameters for region must be re-set for external vectors or bbox. We
can possibly add new columns region_path, region_file for external vectors
and region_bbox with 4 numbers for bbox.
}
\examples{
tb <- scenario_table_create(s_gid = 1,
                            scenset_id = "rondane_ss01",
                            scen_id = "s01",
                            study_area = "Rondane",
                            season = "summer")

# create table with NA data, to be filled later
(tb2 <- scenario_table_create(s_gid = 1:10))

# update table
scenario_table_update(tb2, s_gid = c(3,6), scen_name = "test")

# add new scenarios
scenario_table_add(tb2, s_gid = 11:12, season = "summer", scen_id = paste0("scenario", c(11, 12)))

# error
scenario_table_add(tb2, s_gid = 1)

}
