% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_vect_table.R
\name{scenario_vect_get_values}
\alias{scenario_vect_get_values}
\title{Get attribute table or values from scenario vectors}
\usage{
scenario_vect_get_values(vects, select = NULL)
}
\arguments{
\item{vects}{List of vectors with points, lines, and polygons to be changed in
one or more scenarios.}

\item{select}{Names of columns of the attribute table to be selected/subsetted.
For the column names, see \code{\link[=vect_table_create]{vect_table_create()}}.}
}
\description{
Get attribute table or values from scenario vectors
}
