% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_vect_table.R
\name{scenario_vect_read}
\alias{scenario_vect_read}
\title{Read all vectors from one or all scenarios}
\usage{
scenario_vect_read(
  scenario_table,
  con,
  s_gid = scenario_table$s_gid,
  subset_scenarios = TRUE,
  baseline = TRUE,
  return_format = c("sf", "terra")[2]
)
}
\arguments{
\item{scenario_table}{\verb{[data.frame]} \cr Table of scenarios (as created
through the \code{scenario_table_create()} function).}

\item{con}{\verb{[PqConnection]} \cr Connection to a Postgres database.}

\item{s_gid}{\verb{[numeric(1)]} \cr Scenario general ID - a numeric ID for the
scenario. By default, all scenarios are considered, and all vector are
retrieved.}

\item{subset_scenarios}{\verb{[logical(1)=TRUE]} \cr If \code{TRUE} (default), only the
vector features for the scenario defined by \code{s_gid} are retrieved. If \code{FALSE},
all features from all scenario are read.}

\item{return_format}{\verb{[character(1)="terra"]\{"sf", "terra"\}} \cr Format of the
output vector object in R. So far, only \code{"terra"} is implemented.}

\item{correct_baseline}{\verb{[logical(1)=TRUE]} \cr If \code{TRUE} (default), all
features that should be corrected in the baseline scenario are also retrieved.
This includes all features for which the column \code{changebase} in the vector
attribute table are not \code{NULL}.}
}
\value{
A \code{SpatVector} object.
}
\description{
Read all vectors from one or all scenarios
}
