% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_vect_table.R
\name{vect_table_create}
\alias{vect_table_create}
\title{Create vector layers to be edited for scenario analyses}
\usage{
vect_table_create(
  v_gid = NA_integer_,
  s_gid = NA_integer_,
  scenset_id = NA_character_,
  scen_id = NA_character_,
  type = NA_character_,
  layer = NA_character_,
  param_vals = NA_character_,
  zoi_metric = NA_character_,
  rast_type = NA_character_,
  rast_col = NA_character_,
  zoirad_suit = NA_real_,
  zoirad_perm = NA_real_,
  changebase = c("nochange", "remove", "add", "move")[1],
  change = c("remove", "add", "move")[1],
  repeated = NA_character_,
  comment = NA_character_,
  geom_type = c("POINT", "LINESTRING", "POLYGON", "MULTIPOINT", "MULTILINESTRING",
    "MULTIPOLYGON", "GEOMETRYCOLLECTION")[1],
  geom_col = "geom",
  crs = sf::NA_crs_
)
}
\description{
Create a vector layer with an empty geometry, to be edited in GIS and used to add,
remove, of move features, to be used in scenario analyses.
}
