% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_data.R
\name{db_get_data}
\alias{db_get_data}
\title{Get spatial/grid data from the PostGIS database}
\usage{
db_get_data(
  con,
  table,
  condition = NULL,
  bbox = NULL,
  gid_col = "points_id",
  cols = "*",
  coords = c("x33", "y33"),
  geom = "geom_e33",
  reproject = FALSE,
  crs = 25833,
  query = NULL,
  save_file = ""
)
}
\description{
This function retrieves the spatial grid for Norway annotated with environmental
data. The function is adapted to the version of the reindeer models from
Renewable Reindeer/ProdChange and needs to be changed if model/data structure is
changed.
The function retrieves the spatial grid, including possibly conditions or specific
queries and restricting the data retrieved to a spatial bounding box passed as
a parameter.
}
